requirements.txt
ollama
google-genai
monai
torch
python-docx
pydicom
cryptography
flask
requests
PyMuPDF
google-genai
\nchromadb==0.4.24
